Reorganization of Decoders into Decoders and DataMaps
Split the functionality of Decoders into Decoders and DataMaps

Decoders still represent individual tags, and execute all steps needed to extract their raw data.

DataMaps represent tables in the DB, and describe how the data on disk is to be extracted, transformed, and finally loaded into the DB.

This organization means that an individual Decoder can load multiple tables by implementing multiple DataMaps. This functionality will be useful for those tags which have both instantaneous and summary data. Additionally, this will mean that if/when the DB structure is altered so that data is split into additional tables (i.e. location), writing to these tables will be much easier.

Only Lotek 1000/1100/1250 Decoder is implemented in new pattern. All other decoders are still in the old format and need to be updated.