Update decoders.R
Swapped the order of loading tag metadata and loading tag sensor/summary data to the DB

Means that the call will fail sooner if the tag is already extant in the DB.

Also means that the response from the 'decode' function will now be the number of rows inserted by the final DataMap 'decode_and_load' call. Still not perfect, but better than it was.