Refactor: Tag metadata -> DataMap
The beginning of a refactor I started. The inspiration for this refactor came from the fact that the tag metadata was being treated special out of all of the other data which may come out of a tag. The thought occurred that it may be better practice to instead have the tag metadata be extracted via a separate DataMap object just like all of the other data sources. This would have the benefit of having all data treated the same, and having more code reuse between functions.

The only hitch I encountered was when adding requisite metadata to the regular sensor data before it gets loaded into the DB (i.e. adding the tag_id to the sensor data). Since the tag metadata is just another DataMap now, it necessitates a complex process of going through the DataMaps one by one, checking to see which fields are missing from each, and if those fields are present in other data maps, then copying them over.

Overall, I still have some confidence that this would be the 'better' and more extensible/maintainable implementation. However, at the moment, the benefits it brings are almost entirely theoretical within the scope of this project, and I am struggling to anticipate how these implementations will be directly useful in the future.

As such, I'm mothballing this refactor for now, but keeping these changes as I did make a fair amount of progress.

At this point, I've focused on the Lotek 1000/1100/1250 Decoder/DataMap as a proof of concept. I removed all metadata functionality from the Decoder, and have implemented the tag metadata as its own DataMap. Look in the Lotek script. You'll find my work. I was in the middle of implementing the logic for comparing the fields of each DataMap to the fields of all others. I actually got a breadboard working (lines 48-86 in the lotek.Rmd), but decided at that point that this wasn't worth it.