Slight change in refactor trajectory
After some consideration, decided on a hybrid approach. In this implementation, the tag metadata is in fact handled by a DataMap, but this metadata DataMap occupies a special slot in the Decoder object, so it can be called as necessary.

Actually required implementation of a lot of the logic which would have been required for the original plan, but with a lot less headache.

Only implemented for Lotek 1000/1100/1250 tags so far.