Alternate field names
Refactored the Field object so that it can also specify alternate names for a given field.

This was found to be necessary because a number of Lotek tags which map to the same decoder were found to have different names for the same fields (Temperature & Pressure)

This also necessitated implementing a function which could build POSIXct timestamp objects from multiple possible string formats.