Overwrite check
Decoder objects can now check if the tag ID with which they are associated is already present in the DB.

In addition, TagProcessor and Decoder objects now offer an 'overwrite' option. When set to FALSE, the Decoder will first check if the tag is in the DB, and if it is then the Decoder will not repeat the ETL pipeline.

Handy for saving time on repeat decodings