Lotek processing adjustments
Based on reports of successful/failed decoding, made a number of minor adjustments to the identifiers/decoders/datamaps as deemed neccessary/reasonable.

Adjusted the Datamap for Lotek 1000/1100/1250 tags to only look for the Pressure and Temperature files, without the Light and Supply files. Adjusted the Identifier to do the same.

Changed the Identifier for Lotek 1300 tags to not look for the BIN file, as many tags no longer have it and it's not strictly necessary.