Added field-specific transformation functions
Refactored Field object to offer an optional attribute 'trans_fn' which is applied to just that Field during transformation. trans_fn is applied before all other transformations.

Removes the need for Field to support a number of flags to indicate the need for specific transformations (inversion, casting to numeric, etc).

Refactored the FieldMaps to reflect the new structure