Lotek 1000/1100/1250 bugfix
Fixed a bug where the Lotek 1000-1250 InstantSensor DataMap wouldn't join it's field correctly because for reasons SURPASSING HUMAN UNDERSTANDING the Pressure and Temperature files record their timestamps in different formats.

Not gonna lie, this upsets more than I think is necessarily healthy. It's such a small detail, but it is the rasperry seed in my wisdom tooth that slowly drives me to set my desk on fire.