Populate test DB from build_test_db
Further updated the 'build_test_db' function to allow for a named list of dataframes to be passed in, which will then be used to populate the db.

Updated the DataMap_Base test script to use this new feature. This required removing one of the snapshots, so the snapshot file was also updated.