Removed 'd' field from Decoder class
Removed 'd' as a field from the Decoder class. 'd' is now a parameter passed to the Decoder methods.

This has the benefit of uncoupling the Decoder object from a particular directory, meaning that the same Decoder object can be used to decode the data from multiple directories (produced by the same tag type) without having to construct a separate Decoder object for each directory.