SUBSTANTIAL REFACTOR
Significant refactor of all major objects. DataMaps, Decoders, and Identifiers went from being child implementations of base classes to instantiated objects of the base class type. Instance specific functionality such as the 'extract' function for the DataMap and the 'identify' function for the Identifier were refactored to be lifecycle functions which called other functions specified as fields when the object was instantiated.

This move was made because it became apparent that the current state of the framework, while easier for the developer, was likely too complicated to be used by an average user. Instantiated objects are much more concretely viable for average R users, and easier to implement as well.

This commit is way too big. Really, it should have been split up into a series of commits that tackled the problem piece by piece, but I didn't have time for that, and I was kinda making it up as I went anyways.

All tests currently passing WITH THE NOTABLE EXCEPTION of the TagProcessor class. The TagProcessor class processes the data in the '_test_data' subdirectory, which is real tag data. All of the other tests use mock data. I dunno, it seemed like a good idea at the time. Due to the somewhat complex nature of the DB structure, I found it easier to make a snapshot of an empty DB which can then be copied into temporary directories during testing. However, the snapshot is of the mock data DB, not the real data DB, so the TagProcessor tests always fail.

I've manually determined that the TagProcessor tests do in fact operate correctly, but this should be corrected going forwards, most likely by moving the other tests to use the 'real' data.