Updated FieldMap configurations for timestamps
ABLTAG FieldMaps now convert their POSIXct timestamps into integers (seconds since epoch).

It turns out that its quite difficult to insert DATETIME or TIMESTAMP objects into an Oracle DB from an R environment, but it's easy to insert an integer. So for now, we'll insert an integer and convert it back to a datetime inside of the DB.