TagProcessor dataframe and csv outputs
Renamed process to process_to_db.

Implemented process_to_dataframe and process_to_csv. These perform the exact same function as process_to_db, but return the data as a set of dataframes, or by writing them to csv files in a specified directory.

Renamed build_test_db to build_temp_db and moved it to a util.R file in the main R directory so that it could be accessed by both the testing functions for which it was originally written and the TagProcessor class which now uses it to create a temporary SQLite DB that it loads with the data as normal and then extracts the data from to return as data.frames or csv files.