Removed ABLTAG_TagProcessor
Removed ABLTAG_TagProcessor object in favor of simply moving the default configurations it provided back into the TagProcessor parent object. This way, any users who are not connecting to the ABLTAG DB but still want to take advantage of the pre-configured object can do so from the base object. Additionally, alternative configurations can be passed into the TagProcessor object to adjust its outputs as necessary