Moved base setup commands to util.R
There were a couple of setup commands hanging around in the Field.R file
from when all of the classes were in a single class.R file. Moved them
to the util.R file, which is more of a setup script anyways, and is
therefore loaded before all other scripts.