Field refresh functionality
Added a 'refresh' function to FieldMap and Field class, allowing object
owner of a FieldMap class to quickly pass objects into the Field
objects.

Added so that the TagProcessor, which owns both the output FieldMap
objects and the DB connection object can easily refresh the output
Fields, temporarily providing them with the DB connection object. For
most fields, this is not used, but objects of the InputField_Select class
can use this to refresh the list of labels and ids they present to the
user during data input.

This regular refreshing will be necessary later when we allow the users
to add to the SPECIES table from the input form. When this happens, the
InputField_Select objects will need to refresh their options to reflect
the newly added values