Added augmentation to Decoder processing chain
Decoder now calls 'augment' function as a part of its lifecycle. Augment
function looks for Fields which require user input and then prompts the
user for them.

Updated tests and snaps to reflect this change in the expected data
model. Tests do not use user-input fields, as it would require the user
to input them as the tests ran, slowing testing immeasurably.