InputField tclVar bug
I'm pretty sure I fixed a bug that was lodged in the method by which
InputField objects register new tclVar objects for communicating with
the user input frame. Pretty sure it had to do with pass-by-reference vs
pass-by-value, but to be totally honest I'm not sure. All I know is that
with these changes, it doesn't seem to occur anymore. We'll see if that
holds.