Supressed ODBC message
Added a suppress_messages clause around a persistant error that started
cropping up. Kind of surprised to see it, but it appears to be a known
issue (details in comments of Pipe::upsert method). For now, I can't
figure it out but it isn't breaking anything so I'll leave it for now.