Temporary table cleanup
Upserting into the DB requires copying the data from memory into a
temporary table in the DB. This temporary table is supposed to get
automatically deleted when the connection is severed, but at least in
Oracle this doesn't appear to happen.

Added code which explicitly removes the temporary table immediately
after the upsert operation.