Hanging tables workaround (Oracle)
In order to do UPSERT commands, the package was initially uploading any
new data into a temporary table (table with the TEMPORARY token),
then performing a merge. In theory, tables with the TEMPORARY token
should be automatically deleted after the connection is severed, but
this doesn't actually happen, at least in Oracle. The TEMPORARY tables
also remain 'in use' so long as the connection is open, so they can't
even be deleted by the connection that made them. The only solution was
to close the original connection, which did not delete the tables but
did appear to sever the 'in use' link, and then use a new connection to
delete the TEMPORARY tables. Headache x2.

Turns out that standard tables don't have the 'in use' hangup, so they
can be deleted by the same connection that created them. Changed the
upsert function so that rather than using a TEMPORARY table, the new
data is initially copied into a standard table, merged as normal, then
explicitly deleted at the end of the upsert method call.