Added InputField_SelectAdd
Extended InputField_Select object functionality in new child object
which presents a dropdown of options for the user to choose from, but
which also allows for the user to specify a new value to insert into the
reference table. Also assumes that the table takes care of any PK
creation, etc.