Implemented InputField resetting and persistance
InputFields now get 'reset' by the FieldInputForm when user values get
submitted. At the moment, resetting just involves clearing any and all
tclVar values.

Added a 'persistant' field to the InputField object, which is a boolean
flag indicating if the Field should not be reset between rounds of
inputs. If set to TRUE then the tclVar value will not be cleared on
reset.