Implemented pre-processing step
Pipe now performs a pre-processing step, in which it goes through each
data-directory in the directory tree and performs a pre-processing step
on it without extracting any data.

Currently, this step only includes the identification step, but this
could be extended further into the future. In particular, future
iterations could use this step to build a processing menu.