Added 'data' attribute to directory tree nodes
Nodes on the directory tree now have a 'data' list attribute. Named
elements can be added to this list, and during processing the system
will now check to see if any element in 'data' has the same name as one
of the named output FieldMap objects in the Pipe. If they do, then the
value stored in the 'data' object is used by default. Otherwise the
system processes like normal.