First implementation of FieldWidgets
Added FieldWidget class, which separates the widgets which prompt the
user for input for given fields and the definitions of the Field objects
themselves. This way, a given Field object can be used to spawn multiple
FieldWidget objects, each referring to a separate instance of that
Field (separate tag instance).

This will be necessary for any potential future pre-processing menu that
prompts the user for input for multiple tags simultaneously.

Also implemented a 'data' attribute on the nodes of the directory tree
in the Pipe object. This will allow the system to temporarily store tag
data provided by the user (metadata) while ensuring that each set of
input values remains associated with the corresponding node