Added InputField_FilteredSelect object
Created a new type of InputField object which, like the
InputField_Select object allows the user to choose from a list of
pre-populated options. However, in addition, this new object also allows
the user to make the possible options be filtered by a single
attribute. This is accomplished using the 'filter_field' and 'filter_input_field' attributes.
The filter_field attribute defines the name of the field in the DB table
that will be used to filter the possible choices. In addition to the
field, we also need a value to filter on. This is retrieved using the
input_filter_field attribute, which is a Field object and is used to
isolate the appropriate value.