UPSERT try/catch
Added a try/catch/finally clause to the UPSERT method of the Pipe class.

The upsert method creates a temporary table to load the new data into,
then merges this temporary table with the target table, and finally
drops the temporary table. However, it was possible for an error to be
thrown partway through this process, which would result in the temporary
table never being deleted and being left hanging in the DB.

By wrapping this logic in a try/catch block, I was able to implent the
deletion of the temporary table in the 'finally' clause of the block,
ensuring that no matter what the table will be deleted.