Timestamp bump
Added a function which can bump the timestamps on records which (for one
reason or another) have identical timestamps and therefore can
interfere with the UNIQUE(TAG_ID, TIMESTAMP) constraint in the ABLTAG
tables.

Added the bumping function to the trans_fn of the ABLTAG instant data
output FieldMap so that it will always be invoked prior to such data
being inserted.