Refactored back to manual Decoder
Substantial refactor in the wake of the decision to move pacakage away
from automated processing of large collections of data to targeted
processing of individual data directories.

Originally, the package was designed with the idea of processing large
volumes of backlogged data. However, it was determined that this was not
truly necessary, and that the original inspiration for this
functionality was simply that the developer (me) had been given large
quantities of test data to work with, and had therefore implicitly
assumed that end users would be facing the same challenge.

Pipe class is no longer used, though not yet removed. Metadata such as
tag_id and tag_type are passed into via parameters of the function call.