Unit fixes in the Field configuration file
Fixed a handful of unit references which were using 'G' to refer to G's
(as in gravity) when the `units` package notates that unit more
explicitly as 'gravity'.

Changed the name of the tag # field from 'tag_id' to 'tag_num' to
conform to DB naming scheme

Changed the trans_fn for DesertStar timestamps to include additional
possible formats when converting to POSIXct. These will be necessary for
the two step decoding required for the DesertStar decoders