Reverset fix which caused whole new problem
Users were getting an error which I eventually tracked down to a DBI
function call. Before that though, due to the structure of a tryCatch
clause, it was appearing that the error was occurring in the
dbplyr::db_copy_to function call, so I made a change to the way that the
table names were quoted which I had hoped would fix the problem.

Instead, this created a whole new problem, but because I hadn't figured
out the cause of the initial bug, the two were kind of indistinguishable
from one another.

Eventually, I discovered the problem, and reversed the 'fix'