Removed more vestigial code from InputFields
Removing the user input interface meant that a lot of code became
obsolete, and some it was beoming a problem. Removed the 'augment' step
from the DataMap object and the Decoder method, as there are no longer
any Fields which use this. Also removed the 'refresh' functionality from
both Field and FieldMap objects.

Added a test file for the Field object