Added input Field check
Added a check to the DataMap class which is immediately executed when
the transform process begins. This check ensures that all of the data
fields which are described by the input FieldMap (which are not
independent) are present in the incoming data.

This required me to update the Field objects for a number of pre-defined
Fields, as these fields were not initially present in the input data,
and were instead synthesized from the incoming data later on in the
transform step