Updated testing helper function
Updated function used to test Decoder objects. This function generated
the mock 'input' data from the user, and was missing two fields required
by the output FieldMap.

I believe (though do not know for sure) that this was associated with a
truly bizarre error which would insert the url
"https://cran.rstudio.org" into the outgoing dataframe, or various
variable names from around the package. At this time, I am still baffled
by how these values were making it into the output data, but they were,
and this was my best guess as to how to fix it.