Implemented optional fields
Added an optional flag to the Field object. Default is False, so Fields
are treated as non-optional unless otherwise specified.

Fields in the output FieldMap which are designated as optional will be
ignored by QC checks which ensure that all expected fields are present
in the dataset.

The implementation of this attribute means that the codebase now has to
be able to handle fields which may or may not be populated. Updated
Decoder and DataMap functions to account for this.