Consolidated DataMap file identification
All of the DataMaps use regex patterns to find the files they need,
which opens up the posibility of identifying multiple files when one is
expected.

Implemented a function which wraps the `list.files` function, but
simultaneously checks to see that the number of returned filepaths
matches expectations, and throws an error if they don't.

Updated all the DataMap objects to use this new function