Changed StarOddi DST magnetic FieldMap and Identifier
Updated StarOddi FieldMap to use byte code for degree symbol rather than
the symbol itself, as the symbol was causing problems during
transformation.

Also updated the identifiers for both the DST and the DST magnetic to
ignore field names as they were checking before. This was some leftover
logic from when the package still tried to automatically ID directories
based on their structure