Refactored error reporting in UPSERT method
Because the UPSERT method has a number of internal steps, it was
important to annotate any errors being thrown to make it clear to the
user/developer where the error was being thrown. To do this, I had
implemented a somewhat clunky method of extracting error messages to
construct new error messages which I could then add to to create these
indicators.

The new code is much better, and simply prepends message elements to the
beginning of the message list in the error that is originally thrown.
This has the benefit of being much easier, and also much more effective.

This commit is also accompanied by a change to the DB constraint names.
When a constraint is violated, the constraint name is passed back to the
R code, meaning if the constraint name is meaningful it can inform the
user exactly what is wrong with the data without having to reference the
constraint definition in the DB. I updated all of the constraint names
to describe what they do, so when they get violated it should now be
apparent what is wrong with the outgoing data.