Added DB table metadata to README
Added description attributes to `FieldMap` objects and provided values
for all current `FieldMap`s. Also wrote a function which uses the list
of `FieldMap` objects associated with a given `Decoder` object to
generate a `data.frame` of the metadata for easy display.

Used this function, and the previously implemented function which
generates the same kind of metadata `data.frame` for the individual
fields, in the README file to generate tables that detail the tables and
fields of the ABLTAG database in plain language