WildlifeComputers FieldMap typo
Noticed that the package was generating incorrect UPSERT statements for
the WildlifeComputer MiniPAT tags. Specifically, it was creating MERGE
statements which compared START_TIME and END_TIME fields, which were
should have been left out (START_TIME_POSIXCT and END_TIME_POSIXCT
fields should be the only fields in the MERGE join comparison) and which
also did not include DATA_TYPE and BIN.

Decided to update the FieldMap to explicitly identify the START_TIME and
END_TIME fields as non-id fields, and the problem then resolved.