Updated Field with default value for id_field
Realized that the issue I was seeing was due to the id_field attribute
of the Field class not haveing a default value. This meant that by
default the field had a value of `logical(0)`, which was leading to some
odd results when the FieldMap classes were queried to list their Id
fields. Added the default value to the `initialize` function and the
problem appears to have been resolved.