Updated DataMap and Decoder objects to support optional data sources
Up until now, it has been an implicit requirement that every DataMap
within a Decoder must produce data. However, for some tags this is not
workable, as which types of data are produced can vary.

The motivating example in this case is the WC MiniPAT tag, which can
record Profile of Depth and Temperature (PDT) and Histogram data, but
doesn't always. Both are optional features which are turned on or off by
the user at deployment. As such, some MiniPAT directories will have both
datasets, some will have only one, some will have none. Allowments had
to be made to accomodate this.

Added an 'optional' attribute to the DataMap class, which defaults to
False. When set to True, if an error is caught while executing the
DataMap functions, the error will be displayed as a warning, along with
a statement indicating that the offending DataMap was set to optional,
and so processing will continue. When set to False, the error will be
thrown as normal, and execution will stop.